% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_file.R
\name{process_file}
\alias{process_file}
\title{Procesar un archivo CSV individual (versión robusta)}
\usage{
process_file(input_file, cols_after = 12, keep_table = TRUE)
}
\description{
Procesar un archivo CSV individual (versión robusta)
}
