% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_folder.R
\name{process_folder}
\alias{process_folder}
\title{Procesar carpeta completa}
\usage{
process_folder(input_folder, parallel = TRUE, workers = NULL)
}
\arguments{
\item{input_folder}{ruta a carpeta con CSV}

\item{parallel}{logical usar paralelización}

\item{workers}{número de workers (NULL detecta cores-1)}
}
\description{
Procesar carpeta completa
}
